+++ src/types/piece/queen.rs
@@ -40,8 +40,8 @@ impl Piece for Queen {
         self as &dyn Any
     }
     #[inline(always)]
-    fn score(&self) -> u8 {
-        9
+    fn score(&self) -> f64 {
+        9.
     }
     fn move_set(&self) -> IndexSet<Movement> {
         let rook = Rook::new(self.color, self.pos);
